+++ src/PalphaOrderDependent.cc
@@ -77,7 +77,7 @@ double PalphaOrderDependent::compute (const std::vector< long > & z, int n, doub
             "from that used during object instantiation");
 
    double sum = 0.0;
-   int d = z.size() - 1;
+   long d = z.size() - 1;
 
    for (long k = 0; k < m_n; k++) {
       computeSigma(z, m_C, k);
