+++ src/Writer.cc
@@ -67,7 +67,7 @@ void Writer::writeMScal(const MScal & value)
 
 void Writer::writeMMat(const MMat & A)
 {
-   int sizeA = A.size1();
+   long sizeA = A.size1();
    *m_stream << "   [";
    for (int i = 0; i < sizeA; i++) {
       if (i == 0) { *m_stream << "["; }
